feat: move cli code to internal-folder
