+++ apps/internal-cli/src/commands/config/ai-providers/ai-provider-registry.service.ts
