+++ apps/internal-cli/src/commands/config/show.subcommand.ts
