+++ apps/internal-cli/src/commands/config/test.subcommand.ts
