+++ apps/internal-cli/src/commands/directory/config-check.service.ts
