+++ apps/internal-cli/src/commands/directory/delete.subcommand.ts
