+++ apps/internal-cli/src/commands/directory/directory.command.ts
