+++ apps/internal-cli/src/commands/directory/generate.subcommand.ts
