+++ apps/internal-cli/src/commands/directory/index.ts
