+++ apps/internal-cli/src/commands/directory/list.subcommand.ts
