+++ apps/internal-cli/src/commands/directory/regenerate-markdown.subcommand.ts
