+++ apps/internal-cli/src/commands/directory/remove-item.subcommand.ts
