+++ apps/internal-cli/src/config/config.interface.ts
